fix: use wrapper script for global npm installs
Replace postinstall script with runtime wrapper to fix 'command not found' issue.

Changes:
- Add bin/ck.js wrapper that detects platform and executes correct binary
- Update package.json bin entry from ./bin/ck to ./bin/ck.js
- Remove postinstall script (doesn't run for global installs)
- Update .gitignore to ignore compiled binaries but keep wrapper
- Update GitHub Actions to make wrapper executable
- Add binary-distribution.md documentation

Fixes issue where 'ck' command was not found after global install because
postinstall scripts don't run by default for npm global installs.
